Revert "Preserve LTX2 HQ res2s noise precision"
This reverts commit 01558257b77288dc8ef375a66f83f356ff111752.

+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/ltx_2_denoising.py
@@ -449,12 +449,12 @@ class LTX2DenoisingStage(DenoisingStage):
         midpoint_audio_det = anchor_audio + h * a21 * eps1_audio
 
         sub_noise_video = (
-            self._ltx2_res2s_noise_like(ctx.latents, ctx, substep=True)
+            self._ltx2_res2s_noise_like(ctx.latents, ctx, substep=True).float()
             if ctx.use_native_hq_res2s_sde_noise
             else self._randn_like_with_batch_generators(ctx.latents, batch).float()
         )
         sub_noise_audio = (
-            self._ltx2_res2s_noise_like(ctx.audio_latents, ctx, substep=True)
+            self._ltx2_res2s_noise_like(ctx.audio_latents, ctx, substep=True).float()
             if ctx.use_native_hq_res2s_sde_noise
             else self._randn_like_with_batch_generators(
                 ctx.audio_latents, batch
@@ -523,12 +523,12 @@ class LTX2DenoisingStage(DenoisingStage):
         next_audio_det = anchor_audio + h * (b1 * eps1_audio + b2 * eps2_audio)
 
         step_noise_video = (
-            self._ltx2_res2s_noise_like(ctx.latents, ctx, substep=False)
+            self._ltx2_res2s_noise_like(ctx.latents, ctx, substep=False).float()
             if ctx.use_native_hq_res2s_sde_noise
             else self._randn_like_with_batch_generators(ctx.latents, batch).float()
         )
         step_noise_audio = (
-            self._ltx2_res2s_noise_like(ctx.audio_latents, ctx, substep=False)
+            self._ltx2_res2s_noise_like(ctx.audio_latents, ctx, substep=False).float()
             if ctx.use_native_hq_res2s_sde_noise
             else self._randn_like_with_batch_generators(
                 ctx.audio_latents, batch
@@ -1929,7 +1929,7 @@ class LTX2DenoisingStage(DenoisingStage):
                 midpoint_audio_deterministic = anchor_audio + h * a21 * eps1_audio
 
                 substep_video_noise = (
-                    self._ltx2_res2s_noise_like(ctx.latents, ctx, substep=True)
+                    self._ltx2_res2s_noise_like(ctx.latents, ctx, substep=True).float()
                     if ctx.use_native_hq_res2s_sde_noise
                     else self._randn_like_with_batch_generators(
                         ctx.latents, batch
@@ -1938,7 +1938,7 @@ class LTX2DenoisingStage(DenoisingStage):
                 substep_audio_noise = (
                     self._ltx2_res2s_noise_like(
                         ctx.audio_latents, ctx, substep=True
-                    )
+                    ).float()
                     if ctx.use_native_hq_res2s_sde_noise
                     else self._randn_like_with_batch_generators(
                         ctx.audio_latents, batch
@@ -1998,7 +1998,7 @@ class LTX2DenoisingStage(DenoisingStage):
                 )
 
                 step_video_noise = (
-                    self._ltx2_res2s_noise_like(ctx.latents, ctx, substep=False)
+                    self._ltx2_res2s_noise_like(ctx.latents, ctx, substep=False).float()
                     if ctx.use_native_hq_res2s_sde_noise
                     else self._randn_like_with_batch_generators(
                         ctx.latents, batch
@@ -2007,7 +2007,7 @@ class LTX2DenoisingStage(DenoisingStage):
                 step_audio_noise = (
                     self._ltx2_res2s_noise_like(
                         ctx.audio_latents, ctx, substep=False
-                    )
+                    ).float()
                     if ctx.use_native_hq_res2s_sde_noise
                     else self._randn_like_with_batch_generators(
                         ctx.audio_latents, batch
